requirements.txt
# Core dependencies
fastf1>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0
python-dotenv>=1.0.0

# Data processing
scipy>=1.9.0
scikit-learn>=1.1.0

# Visualization
pillow>=9.0.0
plotly>=5.0.0

# API clients
groq>=0.4.0

# Utilities
click>=8.0.0
rich>=12.0.0
tqdm>=4.64.0

# Development (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.800.0
